requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.3.0
ccxt==4.4.82
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0